chore/deploy and release adjustments
