+++ docker-compose.yml
@@ -57,6 +57,7 @@ services:
       - GeneralSettings:AppLocation=Templates/AspNet/App
       - GeneralSettings:DeploymentLocation=/Templates/deployment
       - GeneralSettings:DeploymentLocation=Templates/AspNet/deployment
+      - GeneralSettings:EnvironmentsUrl=http://studio-azure-mock:6161/environments.json
       - GeneralSettings:IntegrationTestsLocation=Templates/AspNet/App.IntegrationTests
       - GeneralSettings:LanguageFilesLocation=Languages/
       - GeneralSettings:TemplateLocation=/Templates
@@ -72,7 +73,7 @@ services:
       - ServiceRepositorySettings:RepositoryBaseURL=http://studio-repositories:3000
       - ServiceRepositorySettings:RepositoryLocation=/AltinnCore/Repos/
       - TestdataRepositorySettings:RepositoryLocation=/Testdata
-      - Integrations:AzureDevOpsSettings:BaseUri=http://host.docker.internal:6161/_apis/
+      - Integrations:AzureDevOpsSettings:BaseUri=http://studio-azure-mock:6161/_apis/
     ports:
       - "6000:6000"
     depends_on:
