+++ frontend/app-development/index.tsx
@@ -6,6 +6,9 @@ import { setupStore } from './store';
 import { BrowserRouter } from 'react-router-dom';
 import { App } from './App';
 import { APP_DEVELOPMENT_BASENAME } from 'app-shared/constants';
+import { ServicesContextProvider } from './common/ServiceContext';
+import * as queries from './queries/queries';
+import * as mutations from './queries/mutations';
 
 const store = setupStore();
 
@@ -20,7 +23,9 @@ const root = createRoot(container);
 root.render(
   <Provider store={store}>
     <BrowserRouter basename={APP_DEVELOPMENT_BASENAME}>
-      <App />
+      <ServicesContextProvider {...queries} {...mutations}>
+        <App />
+      </ServicesContextProvider>
     </BrowserRouter>
   </Provider>
 );
